refactor(types): split criterion-specific types
